08_MultiAgent/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
PyJWT>=2.8.0
openai>=1.12.0
chromadb>=0.4.22
python-dotenv>=1.0.0
rank-bm25>=0.2.2
sentence-transformers>=2.3.0
tavily-python>=0.3.0
httpx>=0.26.0
python-multipart>=0.0.9
